feat: move changing screen into analytics hook
This allows us to include redux data inside of events. #130
