+++ packages/app/src/components/accounts/index.tsx
@@ -2,12 +2,11 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { Identity } from '@blockstack/keychain';
 import { Text, Flex, FlexProps, Spinner } from '@blockstack/ui';
-
-import { doChangeScreen } from '@store/onboarding/actions';
 import { ScreenName } from '@store/onboarding/types';
 import { PlusInCircle } from '@components/icons/plus-in-circle';
 import { ListItem } from './list-item';
 import { AccountAvatar } from './account-avatar';
+import { useAnalytics } from '@common/hooks/use-analytics';
 
 const loadingProps = { color: '#A1A7B3' };
 const getLoadingProps = (loading: boolean) => (loading ? loadingProps : {});
@@ -53,6 +52,7 @@ interface AccountsProps {
 export const Accounts = ({ identities, showAddAccount, next }: AccountsProps) => {
   const dispatch = useDispatch();
   const [selectedAddress, setSelectedAddress] = useState<null | string>(null);
+  const { doChangeScreen } = useAnalytics();
 
   return (
     <Flex flexDirection="column">
